Add composable face tracker for eyes
