+++ app/src/main/kotlin/org/zxkill/nori/settings/MainSettingsViewModel.kt
@@ -41,4 +41,6 @@ class MainSettingsViewModel @Inject constructor(
     // Устанавливаем время отображения результата скилла
     fun setSkillOutputDisplaySeconds(value: Int) =
         updateData { it.setSkillOutputDisplaySeconds(value) }
+    fun setFaceTrackingDebug(value: Boolean) =
+        updateData { it.setFaceTrackingDebug(value) }
 }
