+++ app/src/main/proto/user_settings.proto
@@ -18,4 +18,6 @@ message UserSettings {
     reserved 9;
     // Время отображения результата скилла на экране в секундах
     int32 skill_output_display_seconds = 10;
+    // Показывать ли отладочное превью камеры рядом с глазами
+    bool face_tracking_debug = 11;
 }
