+++ app/src/main/res/values/strings.xml
@@ -92,6 +92,9 @@
     <string name="pref_weather_default_city_using_ip_info">Set the city to use for weather when you do not explicitly say one. The current behaviour is to get the location from IP info.</string>
     <string name="pref_skill_output_display_time_title">Skill output display time</string>
     <string name="pref_skill_output_display_time_summary">Seconds to show skill result on screen</string>
+    <string name="pref_face_debug_title">Camera debug view</string>
+    <string name="pref_face_debug_summary_on">Camera preview visible</string>
+    <string name="pref_face_debug_summary_off">Camera preview hidden</string>
     <string name="perm_read_contacts">read your contacts</string>
     <string name="perm_call_phone">directly call phone numbers</string>
     <string name="perm_camera">use the camera</string>
@@ -122,10 +125,6 @@
     <string name="skill_sentence_example_current_time">What time is it?</string>
     <string name="skill_name_current_date">Current date</string>
     <string name="skill_sentence_example_current_date">What day is it today?</string>
-    <string name="skill_name_face_tracker">Face tracker</string>
-    <string name="skill_sentence_example_face_tracker">Turn on tracking</string>
-    <string name="skill_face_tracking_enabled">Face tracking enabled</string>
-    <string name="skill_face_tracking_disabled">Face tracking disabled</string>
     <string name="skill_fallback_name_text">Text message</string>
     <string name="skill_search_here_is_what_i_found">Here is what I found</string>
     <string name="skill_search_no_results">The search returned no results, try telling me again what you want to search for</string>
